+++ nodejs/src/delivery/create_entry.ts
@@ -1,13 +1,14 @@
 /* eslint-disable */
 import _m0 from "protobufjs/minimal";
+import { AuthData } from "./auth";
 
 export const protobufPackage = "delivery";
 
 export interface CreateEntryRequest {
   type: string;
-  tenantId: string;
-  data: { [key: string]: string };
+  auth: AuthData | undefined;
   id: string;
+  data: { [key: string]: string };
 }
 
 export interface CreateEntryRequest_DataEntry {
@@ -16,11 +17,23 @@ export interface CreateEntryRequest_DataEntry {
 }
 
 export interface CreateEntryResponse {
-  id: string;
+  newData: { [key: string]: string };
+  validationErrors: string[];
+  fieldValidationErrors: { [key: string]: string };
+}
+
+export interface CreateEntryResponse_NewDataEntry {
+  key: string;
+  value: string;
+}
+
+export interface CreateEntryResponse_FieldValidationErrorsEntry {
+  key: string;
+  value: string;
 }
 
 function createBaseCreateEntryRequest(): CreateEntryRequest {
-  return { type: "", tenantId: "", data: {}, id: "" };
+  return { type: "", auth: undefined, id: "", data: {} };
 }
 
 export const CreateEntryRequest = {
@@ -28,15 +41,15 @@ export const CreateEntryRequest = {
     if (message.type !== "") {
       writer.uint32(10).string(message.type);
     }
-    if (message.tenantId !== "") {
-      writer.uint32(18).string(message.tenantId);
+    if (message.auth !== undefined) {
+      AuthData.encode(message.auth, writer.uint32(18).fork()).ldelim();
+    }
+    if (message.id !== "") {
+      writer.uint32(26).string(message.id);
     }
     Object.entries(message.data).forEach(([key, value]) => {
-      CreateEntryRequest_DataEntry.encode({ key: key as any, value }, writer.uint32(26).fork()).ldelim();
+      CreateEntryRequest_DataEntry.encode({ key: key as any, value }, writer.uint32(34).fork()).ldelim();
     });
-    if (message.id !== "") {
-      writer.uint32(34).string(message.id);
-    }
     return writer;
   },
 
@@ -51,16 +64,16 @@ export const CreateEntryRequest = {
           message.type = reader.string();
           break;
         case 2:
-          message.tenantId = reader.string();
+          message.auth = AuthData.decode(reader, reader.uint32());
           break;
         case 3:
-          const entry3 = CreateEntryRequest_DataEntry.decode(reader, reader.uint32());
-          if (entry3.value !== undefined) {
-            message.data[entry3.key] = entry3.value;
-          }
+          message.id = reader.string();
           break;
         case 4:
-          message.id = reader.string();
+          const entry4 = CreateEntryRequest_DataEntry.decode(reader, reader.uint32());
+          if (entry4.value !== undefined) {
+            message.data[entry4.key] = entry4.value;
+          }
           break;
         default:
           reader.skipType(tag & 7);
@@ -73,28 +86,28 @@ export const CreateEntryRequest = {
   fromJSON(object: any): CreateEntryRequest {
     return {
       type: isSet(object.type) ? String(object.type) : "",
-      tenantId: isSet(object.tenantId) ? String(object.tenantId) : "",
+      auth: isSet(object.auth) ? AuthData.fromJSON(object.auth) : undefined,
+      id: isSet(object.id) ? String(object.id) : "",
       data: isObject(object.data)
         ? Object.entries(object.data).reduce<{ [key: string]: string }>((acc, [key, value]) => {
           acc[key] = String(value);
           return acc;
         }, {})
         : {},
-      id: isSet(object.id) ? String(object.id) : "",
     };
   },
 
   toJSON(message: CreateEntryRequest): unknown {
     const obj: any = {};
     message.type !== undefined && (obj.type = message.type);
-    message.tenantId !== undefined && (obj.tenantId = message.tenantId);
+    message.auth !== undefined && (obj.auth = message.auth ? AuthData.toJSON(message.auth) : undefined);
+    message.id !== undefined && (obj.id = message.id);
     obj.data = {};
     if (message.data) {
       Object.entries(message.data).forEach(([k, v]) => {
         obj.data[k] = v;
       });
     }
-    message.id !== undefined && (obj.id = message.id);
     return obj;
   },
 
@@ -105,14 +118,14 @@ export const CreateEntryRequest = {
   fromPartial(object: DeepPartial<CreateEntryRequest>): CreateEntryRequest {
     const message = createBaseCreateEntryRequest();
     message.type = object.type ?? "";
-    message.tenantId = object.tenantId ?? "";
+    message.auth = (object.auth !== undefined && object.auth !== null) ? AuthData.fromPartial(object.auth) : undefined;
+    message.id = object.id ?? "";
     message.data = Object.entries(object.data ?? {}).reduce<{ [key: string]: string }>((acc, [key, value]) => {
       if (value !== undefined) {
         acc[key] = String(value);
       }
       return acc;
     }, {});
-    message.id = object.id ?? "";
     return message;
   },
 };
@@ -177,14 +190,21 @@ export const CreateEntryRequest_DataEntry = {
 };
 
 function createBaseCreateEntryResponse(): CreateEntryResponse {
-  return { id: "" };
+  return { newData: {}, validationErrors: [], fieldValidationErrors: {} };
 }
 
 export const CreateEntryResponse = {
   encode(message: CreateEntryResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
-    if (message.id !== "") {
-      writer.uint32(10).string(message.id);
+    Object.entries(message.newData).forEach(([key, value]) => {
+      CreateEntryResponse_NewDataEntry.encode({ key: key as any, value }, writer.uint32(10).fork()).ldelim();
+    });
+    for (const v of message.validationErrors) {
+      writer.uint32(18).string(v!);
     }
+    Object.entries(message.fieldValidationErrors).forEach(([key, value]) => {
+      CreateEntryResponse_FieldValidationErrorsEntry.encode({ key: key as any, value }, writer.uint32(26).fork())
+        .ldelim();
+    });
     return writer;
   },
 
@@ -196,7 +216,19 @@ export const CreateEntryResponse = {
       const tag = reader.uint32();
       switch (tag >>> 3) {
         case 1:
-          message.id = reader.string();
+          const entry1 = CreateEntryResponse_NewDataEntry.decode(reader, reader.uint32());
+          if (entry1.value !== undefined) {
+            message.newData[entry1.key] = entry1.value;
+          }
+          break;
+        case 2:
+          message.validationErrors.push(reader.string());
+          break;
+        case 3:
+          const entry3 = CreateEntryResponse_FieldValidationErrorsEntry.decode(reader, reader.uint32());
+          if (entry3.value !== undefined) {
+            message.fieldValidationErrors[entry3.key] = entry3.value;
+          }
           break;
         default:
           reader.skipType(tag & 7);
@@ -207,12 +239,44 @@ export const CreateEntryResponse = {
   },
 
   fromJSON(object: any): CreateEntryResponse {
-    return { id: isSet(object.id) ? String(object.id) : "" };
+    return {
+      newData: isObject(object.newData)
+        ? Object.entries(object.newData).reduce<{ [key: string]: string }>((acc, [key, value]) => {
+          acc[key] = String(value);
+          return acc;
+        }, {})
+        : {},
+      validationErrors: Array.isArray(object?.validationErrors)
+        ? object.validationErrors.map((e: any) => String(e))
+        : [],
+      fieldValidationErrors: isObject(object.fieldValidationErrors)
+        ? Object.entries(object.fieldValidationErrors).reduce<{ [key: string]: string }>((acc, [key, value]) => {
+          acc[key] = String(value);
+          return acc;
+        }, {})
+        : {},
+    };
   },
 
   toJSON(message: CreateEntryResponse): unknown {
     const obj: any = {};
-    message.id !== undefined && (obj.id = message.id);
+    obj.newData = {};
+    if (message.newData) {
+      Object.entries(message.newData).forEach(([k, v]) => {
+        obj.newData[k] = v;
+      });
+    }
+    if (message.validationErrors) {
+      obj.validationErrors = message.validationErrors.map((e) => e);
+    } else {
+      obj.validationErrors = [];
+    }
+    obj.fieldValidationErrors = {};
+    if (message.fieldValidationErrors) {
+      Object.entries(message.fieldValidationErrors).forEach(([k, v]) => {
+        obj.fieldValidationErrors[k] = v;
+      });
+    }
     return obj;
   },
 
@@ -222,7 +286,146 @@ export const CreateEntryResponse = {
 
   fromPartial(object: DeepPartial<CreateEntryResponse>): CreateEntryResponse {
     const message = createBaseCreateEntryResponse();
-    message.id = object.id ?? "";
+    message.newData = Object.entries(object.newData ?? {}).reduce<{ [key: string]: string }>((acc, [key, value]) => {
+      if (value !== undefined) {
+        acc[key] = String(value);
+      }
+      return acc;
+    }, {});
+    message.validationErrors = object.validationErrors?.map((e) => e) || [];
+    message.fieldValidationErrors = Object.entries(object.fieldValidationErrors ?? {}).reduce<
+      { [key: string]: string }
+    >((acc, [key, value]) => {
+      if (value !== undefined) {
+        acc[key] = String(value);
+      }
+      return acc;
+    }, {});
+    return message;
+  },
+};
+
+function createBaseCreateEntryResponse_NewDataEntry(): CreateEntryResponse_NewDataEntry {
+  return { key: "", value: "" };
+}
+
+export const CreateEntryResponse_NewDataEntry = {
+  encode(message: CreateEntryResponse_NewDataEntry, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+    if (message.key !== "") {
+      writer.uint32(10).string(message.key);
+    }
+    if (message.value !== "") {
+      writer.uint32(18).string(message.value);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): CreateEntryResponse_NewDataEntry {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseCreateEntryResponse_NewDataEntry();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.key = reader.string();
+          break;
+        case 2:
+          message.value = reader.string();
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): CreateEntryResponse_NewDataEntry {
+    return { key: isSet(object.key) ? String(object.key) : "", value: isSet(object.value) ? String(object.value) : "" };
+  },
+
+  toJSON(message: CreateEntryResponse_NewDataEntry): unknown {
+    const obj: any = {};
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value);
+    return obj;
+  },
+
+  create(base?: DeepPartial<CreateEntryResponse_NewDataEntry>): CreateEntryResponse_NewDataEntry {
+    return CreateEntryResponse_NewDataEntry.fromPartial(base ?? {});
+  },
+
+  fromPartial(object: DeepPartial<CreateEntryResponse_NewDataEntry>): CreateEntryResponse_NewDataEntry {
+    const message = createBaseCreateEntryResponse_NewDataEntry();
+    message.key = object.key ?? "";
+    message.value = object.value ?? "";
+    return message;
+  },
+};
+
+function createBaseCreateEntryResponse_FieldValidationErrorsEntry(): CreateEntryResponse_FieldValidationErrorsEntry {
+  return { key: "", value: "" };
+}
+
+export const CreateEntryResponse_FieldValidationErrorsEntry = {
+  encode(
+    message: CreateEntryResponse_FieldValidationErrorsEntry,
+    writer: _m0.Writer = _m0.Writer.create(),
+  ): _m0.Writer {
+    if (message.key !== "") {
+      writer.uint32(10).string(message.key);
+    }
+    if (message.value !== "") {
+      writer.uint32(18).string(message.value);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): CreateEntryResponse_FieldValidationErrorsEntry {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseCreateEntryResponse_FieldValidationErrorsEntry();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.key = reader.string();
+          break;
+        case 2:
+          message.value = reader.string();
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): CreateEntryResponse_FieldValidationErrorsEntry {
+    return { key: isSet(object.key) ? String(object.key) : "", value: isSet(object.value) ? String(object.value) : "" };
+  },
+
+  toJSON(message: CreateEntryResponse_FieldValidationErrorsEntry): unknown {
+    const obj: any = {};
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value);
+    return obj;
+  },
+
+  create(
+    base?: DeepPartial<CreateEntryResponse_FieldValidationErrorsEntry>,
+  ): CreateEntryResponse_FieldValidationErrorsEntry {
+    return CreateEntryResponse_FieldValidationErrorsEntry.fromPartial(base ?? {});
+  },
+
+  fromPartial(
+    object: DeepPartial<CreateEntryResponse_FieldValidationErrorsEntry>,
+  ): CreateEntryResponse_FieldValidationErrorsEntry {
+    const message = createBaseCreateEntryResponse_FieldValidationErrorsEntry();
+    message.key = object.key ?? "";
+    message.value = object.value ?? "";
     return message;
   },
 };