+++ nodejs/src/delivery/get_entries.ts
@@ -1,16 +1,24 @@
 /* eslint-disable */
+import Long from "long";
 import _m0 from "protobufjs/minimal";
+import { AuthData } from "./auth";
 import { EntryFilter } from "./entry_filter";
 
 export const protobufPackage = "delivery";
 
-export interface GetEntriesRequest {
+export interface GetEntryRequest {
   type: string;
-  tenantId: string;
+  auth: AuthData | undefined;
   id: string;
+  filter: EntryFilter | undefined;
+  returnEmptyDataIfNotFound: boolean;
+}
+
+export interface GetEntryListRequest {
+  type: string;
+  auth: AuthData | undefined;
   limit: number;
   page: number;
-  returnEmptyDataIfNotFound: boolean;
   filter: EntryFilter | undefined;
   order: EntryOrder[];
 }
@@ -29,44 +37,128 @@ export interface Entry_DataEntry {
   value: string;
 }
 
-export interface GetEntriesResponse {
+export interface GetEntryResponse {
+  result: Entry | undefined;
+}
+
+export interface GetEntryListResponse {
   result: Entry[];
   limit: number;
   page: number;
+  total: number;
 }
 
-function createBaseGetEntriesRequest(): GetEntriesRequest {
-  return {
-    type: "",
-    tenantId: "",
-    id: "",
-    limit: 0,
-    page: 0,
-    returnEmptyDataIfNotFound: false,
-    filter: undefined,
-    order: [],
-  };
+function createBaseGetEntryRequest(): GetEntryRequest {
+  return { type: "", auth: undefined, id: "", filter: undefined, returnEmptyDataIfNotFound: false };
 }
 
-export const GetEntriesRequest = {
-  encode(message: GetEntriesRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+export const GetEntryRequest = {
+  encode(message: GetEntryRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
     if (message.type !== "") {
       writer.uint32(10).string(message.type);
     }
-    if (message.tenantId !== "") {
-      writer.uint32(18).string(message.tenantId);
+    if (message.auth !== undefined) {
+      AuthData.encode(message.auth, writer.uint32(18).fork()).ldelim();
     }
     if (message.id !== "") {
       writer.uint32(26).string(message.id);
     }
-    if (message.limit !== 0) {
-      writer.uint32(32).int32(message.limit);
-    }
-    if (message.page !== 0) {
-      writer.uint32(40).int32(message.page);
+    if (message.filter !== undefined) {
+      EntryFilter.encode(message.filter, writer.uint32(34).fork()).ldelim();
     }
     if (message.returnEmptyDataIfNotFound === true) {
-      writer.uint32(48).bool(message.returnEmptyDataIfNotFound);
+      writer.uint32(40).bool(message.returnEmptyDataIfNotFound);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): GetEntryRequest {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseGetEntryRequest();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.type = reader.string();
+          break;
+        case 2:
+          message.auth = AuthData.decode(reader, reader.uint32());
+          break;
+        case 3:
+          message.id = reader.string();
+          break;
+        case 4:
+          message.filter = EntryFilter.decode(reader, reader.uint32());
+          break;
+        case 5:
+          message.returnEmptyDataIfNotFound = reader.bool();
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): GetEntryRequest {
+    return {
+      type: isSet(object.type) ? String(object.type) : "",
+      auth: isSet(object.auth) ? AuthData.fromJSON(object.auth) : undefined,
+      id: isSet(object.id) ? String(object.id) : "",
+      filter: isSet(object.filter) ? EntryFilter.fromJSON(object.filter) : undefined,
+      returnEmptyDataIfNotFound: isSet(object.returnEmptyDataIfNotFound)
+        ? Boolean(object.returnEmptyDataIfNotFound)
+        : false,
+    };
+  },
+
+  toJSON(message: GetEntryRequest): unknown {
+    const obj: any = {};
+    message.type !== undefined && (obj.type = message.type);
+    message.auth !== undefined && (obj.auth = message.auth ? AuthData.toJSON(message.auth) : undefined);
+    message.id !== undefined && (obj.id = message.id);
+    message.filter !== undefined && (obj.filter = message.filter ? EntryFilter.toJSON(message.filter) : undefined);
+    message.returnEmptyDataIfNotFound !== undefined &&
+      (obj.returnEmptyDataIfNotFound = message.returnEmptyDataIfNotFound);
+    return obj;
+  },
+
+  create(base?: DeepPartial<GetEntryRequest>): GetEntryRequest {
+    return GetEntryRequest.fromPartial(base ?? {});
+  },
+
+  fromPartial(object: DeepPartial<GetEntryRequest>): GetEntryRequest {
+    const message = createBaseGetEntryRequest();
+    message.type = object.type ?? "";
+    message.auth = (object.auth !== undefined && object.auth !== null) ? AuthData.fromPartial(object.auth) : undefined;
+    message.id = object.id ?? "";
+    message.filter = (object.filter !== undefined && object.filter !== null)
+      ? EntryFilter.fromPartial(object.filter)
+      : undefined;
+    message.returnEmptyDataIfNotFound = object.returnEmptyDataIfNotFound ?? false;
+    return message;
+  },
+};
+
+function createBaseGetEntryListRequest(): GetEntryListRequest {
+  return { type: "", auth: undefined, limit: 0, page: 0, filter: undefined, order: [] };
+}
+
+export const GetEntryListRequest = {
+  encode(message: GetEntryListRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+    if (message.type !== "") {
+      writer.uint32(10).string(message.type);
+    }
+    if (message.auth !== undefined) {
+      AuthData.encode(message.auth, writer.uint32(18).fork()).ldelim();
+    }
+    if (message.limit !== 0) {
+      writer.uint32(24).int64(message.limit);
+    }
+    if (message.page !== 0) {
+      writer.uint32(32).int64(message.page);
     }
     if (message.filter !== undefined) {
       EntryFilter.encode(message.filter, writer.uint32(58).fork()).ldelim();
@@ -77,10 +169,10 @@ export const GetEntriesRequest = {
     return writer;
   },
 
-  decode(input: _m0.Reader | Uint8Array, length?: number): GetEntriesRequest {
+  decode(input: _m0.Reader | Uint8Array, length?: number): GetEntryListRequest {
     const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
     let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseGetEntriesRequest();
+    const message = createBaseGetEntryListRequest();
     while (reader.pos < end) {
       const tag = reader.uint32();
       switch (tag >>> 3) {
@@ -88,19 +180,13 @@ export const GetEntriesRequest = {
           message.type = reader.string();
           break;
         case 2:
-          message.tenantId = reader.string();
+          message.auth = AuthData.decode(reader, reader.uint32());
           break;
         case 3:
-          message.id = reader.string();
+          message.limit = longToNumber(reader.int64() as Long);
           break;
         case 4:
-          message.limit = reader.int32();
-          break;
-        case 5:
-          message.page = reader.int32();
-          break;
-        case 6:
-          message.returnEmptyDataIfNotFound = reader.bool();
+          message.page = longToNumber(reader.int64() as Long);
           break;
         case 7:
           message.filter = EntryFilter.decode(reader, reader.uint32());
@@ -116,30 +202,23 @@ export const GetEntriesRequest = {
     return message;
   },
 
-  fromJSON(object: any): GetEntriesRequest {
+  fromJSON(object: any): GetEntryListRequest {
     return {
       type: isSet(object.type) ? String(object.type) : "",
-      tenantId: isSet(object.tenantId) ? String(object.tenantId) : "",
-      id: isSet(object.id) ? String(object.id) : "",
+      auth: isSet(object.auth) ? AuthData.fromJSON(object.auth) : undefined,
       limit: isSet(object.limit) ? Number(object.limit) : 0,
       page: isSet(object.page) ? Number(object.page) : 0,
-      returnEmptyDataIfNotFound: isSet(object.returnEmptyDataIfNotFound)
-        ? Boolean(object.returnEmptyDataIfNotFound)
-        : false,
       filter: isSet(object.filter) ? EntryFilter.fromJSON(object.filter) : undefined,
       order: Array.isArray(object?.order) ? object.order.map((e: any) => EntryOrder.fromJSON(e)) : [],
     };
   },
 
-  toJSON(message: GetEntriesRequest): unknown {
+  toJSON(message: GetEntryListRequest): unknown {
     const obj: any = {};
     message.type !== undefined && (obj.type = message.type);
-    message.tenantId !== undefined && (obj.tenantId = message.tenantId);
-    message.id !== undefined && (obj.id = message.id);
+    message.auth !== undefined && (obj.auth = message.auth ? AuthData.toJSON(message.auth) : undefined);
     message.limit !== undefined && (obj.limit = Math.round(message.limit));
     message.page !== undefined && (obj.page = Math.round(message.page));
-    message.returnEmptyDataIfNotFound !== undefined &&
-      (obj.returnEmptyDataIfNotFound = message.returnEmptyDataIfNotFound);
     message.filter !== undefined && (obj.filter = message.filter ? EntryFilter.toJSON(message.filter) : undefined);
     if (message.order) {
       obj.order = message.order.map((e) => e ? EntryOrder.toJSON(e) : undefined);
@@ -149,18 +228,16 @@ export const GetEntriesRequest = {
     return obj;
   },
 
-  create(base?: DeepPartial<GetEntriesRequest>): GetEntriesRequest {
-    return GetEntriesRequest.fromPartial(base ?? {});
+  create(base?: DeepPartial<GetEntryListRequest>): GetEntryListRequest {
+    return GetEntryListRequest.fromPartial(base ?? {});
   },
 
-  fromPartial(object: DeepPartial<GetEntriesRequest>): GetEntriesRequest {
-    const message = createBaseGetEntriesRequest();
+  fromPartial(object: DeepPartial<GetEntryListRequest>): GetEntryListRequest {
+    const message = createBaseGetEntryListRequest();
     message.type = object.type ?? "";
-    message.tenantId = object.tenantId ?? "";
-    message.id = object.id ?? "";
+    message.auth = (object.auth !== undefined && object.auth !== null) ? AuthData.fromPartial(object.auth) : undefined;
     message.limit = object.limit ?? 0;
     message.page = object.page ?? 0;
-    message.returnEmptyDataIfNotFound = object.returnEmptyDataIfNotFound ?? false;
     message.filter = (object.filter !== undefined && object.filter !== null)
       ? EntryFilter.fromPartial(object.filter)
       : undefined;
@@ -361,39 +438,27 @@ export const Entry_DataEntry = {
   },
 };
 
-function createBaseGetEntriesResponse(): GetEntriesResponse {
-  return { result: [], limit: 0, page: 0 };
+function createBaseGetEntryResponse(): GetEntryResponse {
+  return { result: undefined };
 }
 
-export const GetEntriesResponse = {
-  encode(message: GetEntriesResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
-    for (const v of message.result) {
-      Entry.encode(v!, writer.uint32(10).fork()).ldelim();
-    }
-    if (message.limit !== 0) {
-      writer.uint32(16).int32(message.limit);
-    }
-    if (message.page !== 0) {
-      writer.uint32(24).int32(message.page);
+export const GetEntryResponse = {
+  encode(message: GetEntryResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+    if (message.result !== undefined) {
+      Entry.encode(message.result, writer.uint32(10).fork()).ldelim();
     }
     return writer;
   },
 
-  decode(input: _m0.Reader | Uint8Array, length?: number): GetEntriesResponse {
+  decode(input: _m0.Reader | Uint8Array, length?: number): GetEntryResponse {
     const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
     let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseGetEntriesResponse();
+    const message = createBaseGetEntryResponse();
     while (reader.pos < end) {
       const tag = reader.uint32();
       switch (tag >>> 3) {
         case 1:
-          message.result.push(Entry.decode(reader, reader.uint32()));
-          break;
-        case 2:
-          message.limit = reader.int32();
-          break;
-        case 3:
-          message.page = reader.int32();
+          message.result = Entry.decode(reader, reader.uint32());
           break;
         default:
           reader.skipType(tag & 7);
@@ -403,15 +468,87 @@ export const GetEntriesResponse = {
     return message;
   },
 
-  fromJSON(object: any): GetEntriesResponse {
+  fromJSON(object: any): GetEntryResponse {
+    return { result: isSet(object.result) ? Entry.fromJSON(object.result) : undefined };
+  },
+
+  toJSON(message: GetEntryResponse): unknown {
+    const obj: any = {};
+    message.result !== undefined && (obj.result = message.result ? Entry.toJSON(message.result) : undefined);
+    return obj;
+  },
+
+  create(base?: DeepPartial<GetEntryResponse>): GetEntryResponse {
+    return GetEntryResponse.fromPartial(base ?? {});
+  },
+
+  fromPartial(object: DeepPartial<GetEntryResponse>): GetEntryResponse {
+    const message = createBaseGetEntryResponse();
+    message.result = (object.result !== undefined && object.result !== null)
+      ? Entry.fromPartial(object.result)
+      : undefined;
+    return message;
+  },
+};
+
+function createBaseGetEntryListResponse(): GetEntryListResponse {
+  return { result: [], limit: 0, page: 0, total: 0 };
+}
+
+export const GetEntryListResponse = {
+  encode(message: GetEntryListResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+    for (const v of message.result) {
+      Entry.encode(v!, writer.uint32(10).fork()).ldelim();
+    }
+    if (message.limit !== 0) {
+      writer.uint32(16).int64(message.limit);
+    }
+    if (message.page !== 0) {
+      writer.uint32(24).int64(message.page);
+    }
+    if (message.total !== 0) {
+      writer.uint32(32).int64(message.total);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): GetEntryListResponse {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseGetEntryListResponse();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.result.push(Entry.decode(reader, reader.uint32()));
+          break;
+        case 2:
+          message.limit = longToNumber(reader.int64() as Long);
+          break;
+        case 3:
+          message.page = longToNumber(reader.int64() as Long);
+          break;
+        case 4:
+          message.total = longToNumber(reader.int64() as Long);
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): GetEntryListResponse {
     return {
       result: Array.isArray(object?.result) ? object.result.map((e: any) => Entry.fromJSON(e)) : [],
       limit: isSet(object.limit) ? Number(object.limit) : 0,
       page: isSet(object.page) ? Number(object.page) : 0,
+      total: isSet(object.total) ? Number(object.total) : 0,
     };
   },
 
-  toJSON(message: GetEntriesResponse): unknown {
+  toJSON(message: GetEntryListResponse): unknown {
     const obj: any = {};
     if (message.result) {
       obj.result = message.result.map((e) => e ? Entry.toJSON(e) : undefined);
@@ -420,21 +557,42 @@ export const GetEntriesResponse = {
     }
     message.limit !== undefined && (obj.limit = Math.round(message.limit));
     message.page !== undefined && (obj.page = Math.round(message.page));
+    message.total !== undefined && (obj.total = Math.round(message.total));
     return obj;
   },
 
-  create(base?: DeepPartial<GetEntriesResponse>): GetEntriesResponse {
-    return GetEntriesResponse.fromPartial(base ?? {});
+  create(base?: DeepPartial<GetEntryListResponse>): GetEntryListResponse {
+    return GetEntryListResponse.fromPartial(base ?? {});
   },
 
-  fromPartial(object: DeepPartial<GetEntriesResponse>): GetEntriesResponse {
-    const message = createBaseGetEntriesResponse();
+  fromPartial(object: DeepPartial<GetEntryListResponse>): GetEntryListResponse {
+    const message = createBaseGetEntryListResponse();
     message.result = object.result?.map((e) => Entry.fromPartial(e)) || [];
     message.limit = object.limit ?? 0;
     message.page = object.page ?? 0;
+    message.total = object.total ?? 0;
     return message;
   },
 };
+
+declare var self: any | undefined;
+declare var window: any | undefined;
+declare var global: any | undefined;
+var tsProtoGlobalThis: any = (() => {
+  if (typeof globalThis !== "undefined") {
+    return globalThis;
+  }
+  if (typeof self !== "undefined") {
+    return self;
+  }
+  if (typeof window !== "undefined") {
+    return window;
+  }
+  if (typeof global !== "undefined") {
+    return global;
+  }
+  throw "Unable to locate global object";
+})();
 
 type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;
 
@@ -443,6 +601,18 @@ export type DeepPartial<T> = T extends Builtin ? T
   : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
   : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
   : Partial<T>;
+
+function longToNumber(long: Long): number {
+  if (long.gt(Number.MAX_SAFE_INTEGER)) {
+    throw new tsProtoGlobalThis.Error("Value is larger than Number.MAX_SAFE_INTEGER");
+  }
+  return long.toNumber();
+}
+
+if (_m0.util.Long !== Long) {
+  _m0.util.Long = Long as any;
+  _m0.configure();
+}
 
 function isObject(value: any): boolean {
   return typeof value === "object" && value !== null;