+++ nodejs/src/delivery/update_entry.ts
@@ -1,11 +1,12 @@
 /* eslint-disable */
 import _m0 from "protobufjs/minimal";
+import { AuthData } from "./auth";
 
 export const protobufPackage = "delivery";
 
 export interface UpdateEntryRequest {
   type: string;
-  tenantId: string;
+  auth: AuthData | undefined;
   id: string;
   data: { [key: string]: string };
 }
@@ -16,10 +17,23 @@ export interface UpdateEntryRequest_DataEntry {
 }
 
 export interface UpdateEntryResponse {
+  newData: { [key: string]: string };
+  validationErrors: string[];
+  fieldValidationErrors: { [key: string]: string };
+}
+
+export interface UpdateEntryResponse_NewDataEntry {
+  key: string;
+  value: string;
+}
+
+export interface UpdateEntryResponse_FieldValidationErrorsEntry {
+  key: string;
+  value: string;
 }
 
 function createBaseUpdateEntryRequest(): UpdateEntryRequest {
-  return { type: "", tenantId: "", id: "", data: {} };
+  return { type: "", auth: undefined, id: "", data: {} };
 }
 
 export const UpdateEntryRequest = {
@@ -27,8 +41,8 @@ export const UpdateEntryRequest = {
     if (message.type !== "") {
       writer.uint32(10).string(message.type);
     }
-    if (message.tenantId !== "") {
-      writer.uint32(18).string(message.tenantId);
+    if (message.auth !== undefined) {
+      AuthData.encode(message.auth, writer.uint32(18).fork()).ldelim();
     }
     if (message.id !== "") {
       writer.uint32(26).string(message.id);
@@ -50,7 +64,7 @@ export const UpdateEntryRequest = {
           message.type = reader.string();
           break;
         case 2:
-          message.tenantId = reader.string();
+          message.auth = AuthData.decode(reader, reader.uint32());
           break;
         case 3:
           message.id = reader.string();
@@ -72,7 +86,7 @@ export const UpdateEntryRequest = {
   fromJSON(object: any): UpdateEntryRequest {
     return {
       type: isSet(object.type) ? String(object.type) : "",
-      tenantId: isSet(object.tenantId) ? String(object.tenantId) : "",
+      auth: isSet(object.auth) ? AuthData.fromJSON(object.auth) : undefined,
       id: isSet(object.id) ? String(object.id) : "",
       data: isObject(object.data)
         ? Object.entries(object.data).reduce<{ [key: string]: string }>((acc, [key, value]) => {
@@ -86,7 +100,7 @@ export const UpdateEntryRequest = {
   toJSON(message: UpdateEntryRequest): unknown {
     const obj: any = {};
     message.type !== undefined && (obj.type = message.type);
-    message.tenantId !== undefined && (obj.tenantId = message.tenantId);
+    message.auth !== undefined && (obj.auth = message.auth ? AuthData.toJSON(message.auth) : undefined);
     message.id !== undefined && (obj.id = message.id);
     obj.data = {};
     if (message.data) {
@@ -104,7 +118,7 @@ export const UpdateEntryRequest = {
   fromPartial(object: DeepPartial<UpdateEntryRequest>): UpdateEntryRequest {
     const message = createBaseUpdateEntryRequest();
     message.type = object.type ?? "";
-    message.tenantId = object.tenantId ?? "";
+    message.auth = (object.auth !== undefined && object.auth !== null) ? AuthData.fromPartial(object.auth) : undefined;
     message.id = object.id ?? "";
     message.data = Object.entries(object.data ?? {}).reduce<{ [key: string]: string }>((acc, [key, value]) => {
       if (value !== undefined) {
@@ -176,11 +190,21 @@ export const UpdateEntryRequest_DataEntry = {
 };
 
 function createBaseUpdateEntryResponse(): UpdateEntryResponse {
-  return {};
+  return { newData: {}, validationErrors: [], fieldValidationErrors: {} };
 }
 
 export const UpdateEntryResponse = {
-  encode(_: UpdateEntryResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+  encode(message: UpdateEntryResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+    Object.entries(message.newData).forEach(([key, value]) => {
+      UpdateEntryResponse_NewDataEntry.encode({ key: key as any, value }, writer.uint32(10).fork()).ldelim();
+    });
+    for (const v of message.validationErrors) {
+      writer.uint32(18).string(v!);
+    }
+    Object.entries(message.fieldValidationErrors).forEach(([key, value]) => {
+      UpdateEntryResponse_FieldValidationErrorsEntry.encode({ key: key as any, value }, writer.uint32(26).fork())
+        .ldelim();
+    });
     return writer;
   },
 
@@ -191,6 +215,21 @@ export const UpdateEntryResponse = {
     while (reader.pos < end) {
       const tag = reader.uint32();
       switch (tag >>> 3) {
+        case 1:
+          const entry1 = UpdateEntryResponse_NewDataEntry.decode(reader, reader.uint32());
+          if (entry1.value !== undefined) {
+            message.newData[entry1.key] = entry1.value;
+          }
+          break;
+        case 2:
+          message.validationErrors.push(reader.string());
+          break;
+        case 3:
+          const entry3 = UpdateEntryResponse_FieldValidationErrorsEntry.decode(reader, reader.uint32());
+          if (entry3.value !== undefined) {
+            message.fieldValidationErrors[entry3.key] = entry3.value;
+          }
+          break;
         default:
           reader.skipType(tag & 7);
           break;
@@ -199,12 +238,45 @@ export const UpdateEntryResponse = {
     return message;
   },
 
-  fromJSON(_: any): UpdateEntryResponse {
-    return {};
+  fromJSON(object: any): UpdateEntryResponse {
+    return {
+      newData: isObject(object.newData)
+        ? Object.entries(object.newData).reduce<{ [key: string]: string }>((acc, [key, value]) => {
+          acc[key] = String(value);
+          return acc;
+        }, {})
+        : {},
+      validationErrors: Array.isArray(object?.validationErrors)
+        ? object.validationErrors.map((e: any) => String(e))
+        : [],
+      fieldValidationErrors: isObject(object.fieldValidationErrors)
+        ? Object.entries(object.fieldValidationErrors).reduce<{ [key: string]: string }>((acc, [key, value]) => {
+          acc[key] = String(value);
+          return acc;
+        }, {})
+        : {},
+    };
   },
 
-  toJSON(_: UpdateEntryResponse): unknown {
+  toJSON(message: UpdateEntryResponse): unknown {
     const obj: any = {};
+    obj.newData = {};
+    if (message.newData) {
+      Object.entries(message.newData).forEach(([k, v]) => {
+        obj.newData[k] = v;
+      });
+    }
+    if (message.validationErrors) {
+      obj.validationErrors = message.validationErrors.map((e) => e);
+    } else {
+      obj.validationErrors = [];
+    }
+    obj.fieldValidationErrors = {};
+    if (message.fieldValidationErrors) {
+      Object.entries(message.fieldValidationErrors).forEach(([k, v]) => {
+        obj.fieldValidationErrors[k] = v;
+      });
+    }
     return obj;
   },
 
@@ -212,8 +284,148 @@ export const UpdateEntryResponse = {
     return UpdateEntryResponse.fromPartial(base ?? {});
   },
 
-  fromPartial(_: DeepPartial<UpdateEntryResponse>): UpdateEntryResponse {
+  fromPartial(object: DeepPartial<UpdateEntryResponse>): UpdateEntryResponse {
     const message = createBaseUpdateEntryResponse();
+    message.newData = Object.entries(object.newData ?? {}).reduce<{ [key: string]: string }>((acc, [key, value]) => {
+      if (value !== undefined) {
+        acc[key] = String(value);
+      }
+      return acc;
+    }, {});
+    message.validationErrors = object.validationErrors?.map((e) => e) || [];
+    message.fieldValidationErrors = Object.entries(object.fieldValidationErrors ?? {}).reduce<
+      { [key: string]: string }
+    >((acc, [key, value]) => {
+      if (value !== undefined) {
+        acc[key] = String(value);
+      }
+      return acc;
+    }, {});
+    return message;
+  },
+};
+
+function createBaseUpdateEntryResponse_NewDataEntry(): UpdateEntryResponse_NewDataEntry {
+  return { key: "", value: "" };
+}
+
+export const UpdateEntryResponse_NewDataEntry = {
+  encode(message: UpdateEntryResponse_NewDataEntry, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+    if (message.key !== "") {
+      writer.uint32(10).string(message.key);
+    }
+    if (message.value !== "") {
+      writer.uint32(18).string(message.value);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): UpdateEntryResponse_NewDataEntry {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseUpdateEntryResponse_NewDataEntry();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.key = reader.string();
+          break;
+        case 2:
+          message.value = reader.string();
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): UpdateEntryResponse_NewDataEntry {
+    return { key: isSet(object.key) ? String(object.key) : "", value: isSet(object.value) ? String(object.value) : "" };
+  },
+
+  toJSON(message: UpdateEntryResponse_NewDataEntry): unknown {
+    const obj: any = {};
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value);
+    return obj;
+  },
+
+  create(base?: DeepPartial<UpdateEntryResponse_NewDataEntry>): UpdateEntryResponse_NewDataEntry {
+    return UpdateEntryResponse_NewDataEntry.fromPartial(base ?? {});
+  },
+
+  fromPartial(object: DeepPartial<UpdateEntryResponse_NewDataEntry>): UpdateEntryResponse_NewDataEntry {
+    const message = createBaseUpdateEntryResponse_NewDataEntry();
+    message.key = object.key ?? "";
+    message.value = object.value ?? "";
+    return message;
+  },
+};
+
+function createBaseUpdateEntryResponse_FieldValidationErrorsEntry(): UpdateEntryResponse_FieldValidationErrorsEntry {
+  return { key: "", value: "" };
+}
+
+export const UpdateEntryResponse_FieldValidationErrorsEntry = {
+  encode(
+    message: UpdateEntryResponse_FieldValidationErrorsEntry,
+    writer: _m0.Writer = _m0.Writer.create(),
+  ): _m0.Writer {
+    if (message.key !== "") {
+      writer.uint32(10).string(message.key);
+    }
+    if (message.value !== "") {
+      writer.uint32(18).string(message.value);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): UpdateEntryResponse_FieldValidationErrorsEntry {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseUpdateEntryResponse_FieldValidationErrorsEntry();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.key = reader.string();
+          break;
+        case 2:
+          message.value = reader.string();
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): UpdateEntryResponse_FieldValidationErrorsEntry {
+    return { key: isSet(object.key) ? String(object.key) : "", value: isSet(object.value) ? String(object.value) : "" };
+  },
+
+  toJSON(message: UpdateEntryResponse_FieldValidationErrorsEntry): unknown {
+    const obj: any = {};
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value);
+    return obj;
+  },
+
+  create(
+    base?: DeepPartial<UpdateEntryResponse_FieldValidationErrorsEntry>,
+  ): UpdateEntryResponse_FieldValidationErrorsEntry {
+    return UpdateEntryResponse_FieldValidationErrorsEntry.fromPartial(base ?? {});
+  },
+
+  fromPartial(
+    object: DeepPartial<UpdateEntryResponse_FieldValidationErrorsEntry>,
+  ): UpdateEntryResponse_FieldValidationErrorsEntry {
+    const message = createBaseUpdateEntryResponse_FieldValidationErrorsEntry();
+    message.key = object.key ?? "";
+    message.value = object.value ?? "";
     return message;
   },
 };